refactor: split logic to separate files, delete console.log
