+++ client/src/components/ConfirmPage/Container/style.sass
@@ -1,0 +1,11 @@
+.confirm-page
+
+  &__container
+    display: flex
+    justify-content: space-between
+    align-items: flex-start
+    width: 1600px
+    margin: 0 auto
+
+  &__image
+    width: 45%
