+++ client/src/components/ConfirmPage/Info/style.sass
@@ -2,20 +2,12 @@
 
 .confirm-page
 
-  &__container
-    display: flex
-    align-items: flex-start
-    width: 1600px
-    margin: 0 auto
-
-  &__image
-    width: 45%
-
   &__info
     width: 45%
     margin-top: 200px
+    font-size: 18px
     text-align: center
-    color: #7084c5
+    color: #54a6dd
 
   &__text
     margin-top: 20px
